proyecto/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
cryptography==44.0.2
cssselect==1.2.0
cssutils==2.11.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
mysql-connector-python==9.2.0
packaging==24.2
premailer==3.10.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yagmail==0.15.293
